グラフを用いて直近一週間の書いたメモの数を表示できるようにした。

+++ Achievement/Controller/Graph/GraphData.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct JournalsCount {
     var showNumber: String
-    var viewCount:Double
+    var viewCount:Int
 }
